Move Monitor into @miniplex/bucket
