+++ packages/miniplex-bucket/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./Bucket"
+export * from "./Monitor"
